requirements.txt
gradio
transformers
torch
Pillow
opencv-python
sentencepiece